Implemented and compiled, but not tested, mesh creation and propagation to subsidiary pools etc. Replaced usage of printf in MatrixOps.cpp with cout.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/threadMsg@2816 d1858a74-f715-0410-99c9-c66fe9cce3d6

+++ regressionTests/rtHHnetwork.cpp
@@ -333,7 +333,9 @@ void rtHHnetwork( unsigned int numCopies )
 	shell->doSyncDataHandler( synCopyId );
 
 	// Make it twice as big as expected probability, for safety.
-	unsigned int numConnections = 2 * numCopies * sqrt( numCopies );
+	unsigned int numConnections = 
+		2 * numCopies * static_cast< unsigned int >( 
+		sqrt( double( numCopies ) ) );
 	mtseed( 1000UL );
 	vector< double > weight( numConnections );
 	vector< double > delay( numConnections );
